+++ src/parser/type.cpp
@@ -31,7 +31,7 @@ Node::Type *Parser::symbol_table(PStruct *psr) {
 Node::Type *Parser::array_type(PStruct *psr) {
   psr->advance(psr);
   // Check if the next token is an integer (const size)
-  signed short int size = 0; // 0 is default, uninitialized
+  size_t size = 0; // 0 is default, uninitialized
   if (psr->peek(psr).kind == TokenKind::INT) {
     // There will be warnings and possible
     // overflow because stoi returns a 32-bit
